reorganize fees traits
